pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Leonardo Lazzaro <llazzaro@dc.uba.ar>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
psycopg = "^3.1.18"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"
djlint = "^1.34.1"
djangorestframework = "^3.14.0"
django-storages = "^1.14.2"
markdown = "^3.5.2"
django-widget-tweaks = "^1.5.0"
pillow = "^10.2.0"
django-ninja = "^1.1.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
pyjwt = "^2.8.0"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.djlint]
profile = "django"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "taskmanager.settings"
